Fix various record properties such as is_deletion
Add some properties to the record (is_transversion and is_insertion) for
completion and fix the implementation of some others. The is_deletion
property is no longer True for monomorphic records and insertions and
deletions are only called after checking every alternative allele. The
code is repetitive and not pretty, it could use some refactoring.